switch operation details to a drawer
