+++ src/renderer/drawers/OperationDetails/styledComponents.js
