+++ src/renderer/families/polkadot/UnbondFlowModal/steps/StepConfirmation.js
@@ -12,6 +12,8 @@ import RetryButton from "~/renderer/components/RetryButton";
 import ErrorDisplay from "~/renderer/components/ErrorDisplay";
 import SuccessDisplay from "~/renderer/components/SuccessDisplay";
 import BroadcastErrorDisclaimer from "~/renderer/components/BroadcastErrorDisclaimer";
+import { OperationDetails } from "~/renderer/drawers/OperationDetails";
+import { setDrawer } from "~/renderer/drawers/Provider";
 
 import type { StepProps } from "../types";
 
@@ -88,7 +90,7 @@ export function StepConfirmationFooter({
           onClick={() => {
             onClose();
             if (account && optimisticOperation) {
-              openModal("MODAL_OPERATION_DETAILS", {
+              setDrawer(OperationDetails, {
                 operationId: optimisticOperation.id,
                 accountId: account.id,
                 parentId: parentAccount && parentAccount.id,
