+++ src/renderer/families/tezos/DelegateFlowModal/steps/StepConfirmation.js
@@ -1,12 +1,10 @@
 // @flow
 import invariant from "invariant";
 import React from "react";
-import { useDispatch } from "react-redux";
 import styled, { withTheme } from "styled-components";
 import { Trans } from "react-i18next";
 import { SyncOneAccountOnMount } from "@ledgerhq/live-common/lib/bridge/react";
 import { multiline } from "~/renderer/styles/helpers";
-import { openModal } from "~/renderer/actions/modals";
 import { urls } from "~/config/urls";
 import TrackPage from "~/renderer/analytics/TrackPage";
 import Box from "~/renderer/components/Box";
@@ -18,6 +16,8 @@ import type { ThemedComponent } from "~/renderer/styles/StyleProvider";
 import ErrorDisplay from "~/renderer/components/ErrorDisplay";
 import SuccessDisplay from "~/renderer/components/SuccessDisplay";
 import BroadcastErrorDisclaimer from "~/renderer/components/BroadcastErrorDisclaimer";
+import { OperationDetails } from "~/renderer/drawers/OperationDetails";
+import { setDrawer } from "~/renderer/drawers/Provider";
 
 import type { StepProps } from "../types";
 
@@ -101,7 +101,6 @@ export const StepConfirmationFooter = ({
   error,
   closeModal,
 }: StepProps) => {
-  const dispatch = useDispatch();
   const concernedOperation = optimisticOperation
     ? optimisticOperation.subOperations && optimisticOperation.subOperations.length > 0
       ? optimisticOperation.subOperations[0]
@@ -123,13 +122,11 @@ export const StepConfirmationFooter = ({
           onClick={() => {
             closeModal();
             if (account && concernedOperation) {
-              dispatch(
-                openModal("MODAL_OPERATION_DETAILS", {
-                  operationId: concernedOperation.id,
-                  accountId: account.id,
-                  parentId: parentAccount && parentAccount.id,
-                }),
-              );
+              setDrawer(OperationDetails, {
+                operationId: concernedOperation.id,
+                accountId: account.id,
+                parentId: parentAccount && parentAccount.id,
+              });
             }
           }}
           primary
