+++ src/renderer/modals/Swap/steps/StepFinished.js
@@ -10,12 +10,12 @@ import Text from "~/renderer/components/Text";
 import styled from "styled-components";
 import { colors } from "~/renderer/styles/theme";
 import Button from "~/renderer/components/Button";
-import { openModal } from "~/renderer/actions/modals";
 import InfoCircle from "~/renderer/icons/InfoCircle";
 import useTheme from "~/renderer/hooks/useTheme";
-import { useDispatch } from "react-redux";
-import { GradientHover } from "~/renderer/modals/OperationDetails/styledComponents";
+import { GradientHover } from "~/renderer/drawers/OperationDetails/styledComponents";
 import type { ThemedComponent } from "~/renderer/styles/StyleProvider";
+import { OperationDetails } from "~/renderer/drawers/OperationDetails";
+import { setDrawer } from "~/renderer/drawers/Provider";
 
 const IconWrapper = styled(Box)`
   background: ${colors.pillActiveBackground};
@@ -139,7 +139,6 @@ export const StepFinishedFooter = ({
 }) => {
   const { operation } = result;
   const { fromAccount, fromParentAccount } = swap.exchange;
-  const dispatch = useDispatch();
 
   const onViewOperationDetails = useCallback(() => {
     const concernedOperation = operation
@@ -150,15 +149,13 @@ export const StepFinishedFooter = ({
 
     onClose();
     if (fromAccount && concernedOperation) {
-      dispatch(
-        openModal("MODAL_OPERATION_DETAILS", {
-          operationId: concernedOperation.id,
-          accountId: fromAccount.id,
-          parentId: fromParentAccount && fromParentAccount.id,
-        }),
-      );
+      setDrawer(OperationDetails, {
+        operationId: concernedOperation.id,
+        accountId: fromAccount.id,
+        parentId: fromParentAccount && fromParentAccount.id,
+      });
     }
-  }, [dispatch, fromAccount, fromParentAccount, onClose, operation]);
+  }, [fromAccount, fromParentAccount, onClose, operation]);
 
   return (
     <Box horizontal>
