+++ src/renderer/modals/index.js
@@ -12,7 +12,6 @@ import MODAL_RECEIVE from "./Receive";
 import MODAL_TERMS from "./Terms";
 import MODAL_SEND from "./Send";
 import MODAL_UPDATE_FIRMWARE from "./UpdateFirmwareModal";
-import MODAL_OPERATION_DETAILS from "./OperationDetails";
 import MODAL_MIGRATE_ACCOUNTS from "./MigrateAccounts";
 import MODAL_EXPORT_ACCOUNTS from "./ExportAccounts";
 import MODAL_TECHNICAL_DATA from "./TechnicalData";
@@ -73,7 +72,6 @@ const modals: { [_: string]: React$ComponentType<any> } = {
   MODAL_TERMS,
   MODAL_SEND,
   MODAL_UPDATE_FIRMWARE,
-  MODAL_OPERATION_DETAILS,
   MODAL_DELEGATE,
   MODAL_MIGRATE_ACCOUNTS,
   MODAL_EXPORT_ACCOUNTS,
